adding redirecting grid , animated button & plans page

+++ lib/API/myRequestsModel.dart
@@ -54,7 +54,7 @@ class RequestElement {
 class Plan {
   String description;
   List<String> options;
-  int cost;
+  double cost;
   int color;
   String id;
   String name;
@@ -79,7 +79,7 @@ class Plan {
   factory Plan.fromJson(Map<String, dynamic> json) => Plan(
     description: json["description"],
     options: List<String>.from(json["options"].map((x) => x)),
-    cost: json["cost"],
+    cost: json["cost"].toDouble(),
     color: json["color"],
     id: json["_id"],
     name: json["name"],
